FaceAnalysisApp/requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.12
dlib==19.19.0
gast==0.2.2
gcloud==0.17.0
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
h5py==2.10.0
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
imutils==0.5.3
joblib==0.14.1
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
Markdown==3.2.2
mysqlclient==1.4.6
numpy==1.18.4
oauth2client==3.0.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-firebase==1.2
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.20.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0